Reorganize Core and Standard files, update includes
