+++ src/Core/Engine/Engine.cpp
@@ -1,4 +1,4 @@
-#include "WorldInstance.h"
+#include "Engine.h"
 
 namespace CGEngine {
     //Size and name to give created window
