+++ src/Core/Logging/Logging.cpp
