+++ src/Core/Logging/Logging.h
