+++ src/Core/ResourceCaches/FontCache.cpp
