+++ src/Core/ResourceCaches/TextureCache.cpp
