+++ src/Core/ResourceCaches/TextureCache.h
