+++ src/Core/Types/UniqueIntegerStack.h
