+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "World/WorldInstance.h"
+#include "Core/Engine/Engine.h"
 using namespace CGEngine;
 
 int main() {
